Create ImportOlympiaNameLedgerAndDerivePublicKeys file with current implementation
